Use proper integer types, continue cleanup

+++ src/plan_tree_walker.h
@@ -37,11 +37,11 @@ extern void pwh_walk_planstate_tree(PlanState		*planstate,
 									PwhNodeVisitorFn visitor, void *context);
 
 /* Walk plan tree and populate topology (node_id, parent_id, node_type). */
-extern usize pwh_walk_plan_topology(PlanState *planstate, PwhNode *metrics,
-									usize max_nodes, i32 parent_id);
+extern u64 pwh_walk_plan_topology(PlanState *planstate, PwhNode *metrics,
+								  u64 max_nodes, i32 parent_id);
 
 /* Walk plan tree and read Instrumentation data. */
 extern void pwh_walk_plan_instrumentation(PlanState *planstate,
-										  PwhNode *metrics, usize max_nodes);
+										  PwhNode *metrics, u64 max_nodes);
 
 #endif /* PWH_PLAN_TREE_WALKER_H. */
